feat: break and continue statements
feat(test): added more test scripts

feat: 16 bit constants

feat: stack overflow safety

chore: cleanup
