+++ src/chunk.h
@@ -83,6 +83,17 @@ typedef enum {
   OP_USE_NATIVE,
   OP_USE_MODULE,
   OP_FINISH_USE,
+  OP_CONSTANT_16,
+  OP_DEFINE_GLOBAL_16,
+  OP_GET_GLOBAL_16,
+  OP_SET_GLOBAL_16,
+  OP_GET_PROPERTY_16,
+  OP_SET_PROPERTY_16,
+  OP_INVOKE_16,
+  OP_SUPER_INVOKE_16,
+  OP_GET_SUPER_16,
+  OP_CLASS_16,
+  OP_METHOD_16
 } OpCode;
 
 typedef struct {
