+++ src/common.h
@@ -10,10 +10,11 @@
 // #define DEBUG_LOG_GC
 // #define DEBUG_STRESS_GC
 // #define DUMP_BYTECODE
+
 #define UINT8_COUNT (UINT8_MAX + 1)
 #define MAX_ARRAY_SIZE UINT16_MAX - 1
 #define FRAMES_MAX 64
-#define STACK_MAX (FRAMES_MAX * UINT8_COUNT)
+#define STACK_MAX (FRAMES_MAX * UINT8_COUNT * 8) // Approx 1MB stack size
 #define IMPORT_MAX FRAMES_MAX / 2
 
 #endif
