+++ src/main.c
@@ -2,7 +2,7 @@
 #include <stdlib.h>
 
 #include "file_handler.h"
-#include "vm.h"
+#include "vm/vm.h"
 
 /**
  * Starts an interactive Read-Eval-Print Loop (REPL) for the language
